blog-backend/requirement.txt
amqp==5.1.1
anyascii==0.3.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.1.9
django-allauth==0.51.0
django-appconf==1.0.5
django-celery-beat==2.4.0
django-compressor==4.4
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
django-filter==23.4
django-modelcluster==6.1
django-on-heroku==1.1.2
django-pagination-bootstrap==2.4.6
django-permissionedforms==0.1
django-phonenumber-field==7.0.0
django-rest-passwordreset==1.3.0
django-taggit==4.0.0
django-timezone-field==5.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filelock==3.7.1
filetype==1.2.0
google-api-python-client==1.6.4
html5lib==1.1
httplib2==0.20.4
idna==3.4
importlib-metadata==5.2.0
jsonpickle==0.9.6
kombu==5.2.4
l18n==2021.3
Markdown==3.4.1
oauth2client==4.1.2
oauthlib==3.2.1
openpyxl==3.1.2
phonenumberslite==8.12.54
Pillow==9.2.0
pillow-heif==0.13.1
pipenv==2022.6.7
platformdirs==2.5.2
prompt-toolkit==3.0.36
psycopg2==2.9.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
rcssmin==1.1.1
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
telepath==0.3.1
tzdata==2022.2
uritemplate==3.0.1
urllib3==1.26.13
vboxapi==1.0
vine==5.0.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wagtail==5.2.1
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
Willow==1.6.3
zipp==3.11.0
